incunabulaBackend/requirements.txt
asgiref==3.3.4
django==3.2.20
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
h5py==3.1.0
keras==2.2.4
keras-applications==1.0.8
keras-preprocessing==1.1.2
numpy==1.19.5
pillow==8.0.0
psycopg2-binary==2.9.8
pytz==2024.1
pyyaml==6.0
scipy==1.5.4
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.7.4
rdflib==5.0.0
python-decouple==3.8
